requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bokeh==3.4.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dash==2.17.0
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dbg==1.5.2
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.51.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
imbalanced-learn==0.12.2
imblearn==0.0
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
openai==1.30.1
orjson==3.10.3
packaging==24.0
pandas==2.2.2
parso==0.8.4
patsy==0.5.6
pillow==10.3.0
platformdirs==4.2.2
plotly==5.22.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pydantic==2.7.1
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
pymongo==4.7.2
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.31.0
retrying==1.3.4
rich==13.7.1
scikit-learn==1.4.2
scipy==1.13.0
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.2
tenacity==8.3.0
threadpoolctl==3.5.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
wordcloud==1.9.3
xgboost==2.0.3
xyzservices==2024.4.0
yarl==1.9.4
zipp==3.18.2
